clip more JS codes on native (#7520)
