+++ cocos/core/3d/framework/skinned-mesh-batch-renderer.ts
@@ -36,9 +36,8 @@ import { Skeleton } from '../../assets/skeleton';
 import { Texture2D } from '../../assets/texture-2d';
 import { ccclass, help, executeInEditMode, executionOrder, menu, tooltip, type, visible, override, serializable, editable } from 'cc.decorator';
 import { CCString } from '../../data/utils/attribute';
-import { GFXAttributeName, GFXBufferTextureCopy, GFXFormatInfos } from '../../gfx/define';
-import { GFXFormat, GFXType } from '../../gfx/define';
-import { GFXDevice, GFXAttribute } from '../../gfx';
+import { GFXAttributeName, GFXFormatInfos, GFXFormat, GFXType } from '../../gfx/define';
+import { GFXDevice, GFXAttribute, GFXBufferTextureCopy } from '../../gfx';
 import { Mat4, Vec2, Vec3 } from '../../math';
 import { mapBuffer, readBuffer, writeBuffer } from '../misc/buffer';
 import { SkinnedMeshRenderer } from './skinned-mesh-renderer';
