+++ cocos/core/gfx/device.ts
@@ -7,7 +7,8 @@ import { ccenum } from '../value-types/enum';
 import { GFXDescriptorSet, GFXDescriptorSetInfo } from './descriptor-set';
 import { GFXBuffer, GFXBufferInfo, GFXBufferViewInfo } from './buffer';
 import { GFXCommandBuffer, GFXCommandBufferInfo } from './command-buffer';
-import { GFXBufferTextureCopy, GFXFilter, GFXFormat, GFXMemoryStatus, GFXRect, GFXAPI, GFXFeature } from './define';
+import {  GFXFilter, GFXFormat, GFXMemoryStatus, GFXAPI, GFXFeature } from './define';
+import { GFXBufferTextureCopy, GFXRect } from './define-class';
 import { GFXFence, GFXFenceInfo } from './fence';
 import { GFXFramebuffer, GFXFramebufferInfo } from './framebuffer';
 import { GFXInputAssembler, GFXInputAssemblerInfo } from './input-assembler';
