+++ cocos/core/gfx/webgl/webgl-state-cache.ts
@@ -1,4 +1,4 @@
-import { GFXRect, GFXViewport } from '../define';
+import { GFXRect, GFXViewport } from '../define-class';
 import { GFXBlendState, GFXDepthStencilState, GFXRasterizerState } from '../pipeline-state';
 
 export interface IWebGLTexUnit {
