+++ cocos/core/pipeline/shadow/shadow-stage.ts
@@ -4,7 +4,7 @@
  */
 
 import { ccclass } from 'cc.decorator';
-import { GFXColor, GFXRect } from '../../gfx/define';
+import { GFXColor, GFXRect } from '../../gfx';
 import { IRenderStageInfo, RenderStage } from '../render-stage';
 import { RenderView } from '../render-view';
 import { ForwardStagePriority } from '../forward/enum';
