+++ cocos/core/renderer/core/pass.ts
@@ -34,7 +34,6 @@ import { IPassInfo, IPassStates, IPropertyInfo } from '../../assets/effect-asset
 import { TextureBase } from '../../assets/texture-base';
 import { GFXDescriptorSet, GFXDescriptorSetInfo } from '../../gfx/descriptor-set';
 import { GFXBlendState, GFXBlendTarget, GFXDepthStencilState, GFXRasterizerState } from '../../gfx/pipeline-state';
-import { GFXSampler } from '../../gfx/sampler';
 import { RenderPassStage, RenderPriority, SetIndex } from '../../pipeline/define';
 import { getPhaseID } from '../../pipeline/pass-phase';
 import { Root } from '../../root';
@@ -47,7 +46,7 @@ import { customizeType, getBindingFromHandle, getPropertyTypeFromHandle, getDefa
     getOffsetFromHandle, getTypeFromHandle, MacroRecord, MaterialProperty, type2reader, type2writer, PropertyType } from './pass-utils';
 import { GFXBufferUsageBit, GFXGetTypeSize, GFXMemoryUsageBit, GFXPrimitiveMode,
     GFXType, GFXDynamicStateFlagBit, GFXDynamicStateFlags, GFXFeature } from '../../gfx/define';
-import { GFXDescriptorSetLayoutInfo, GFXTexture,  GFXDevice, GFXBuffer, GFXBufferInfo, GFXBufferViewInfo } from '../../gfx';
+import { GFXDescriptorSetLayoutInfo, GFXTexture,  GFXDevice, GFXBuffer, GFXBufferInfo, GFXBufferViewInfo, GFXSampler } from '../../gfx';
 
 export interface IPassInfoFull extends IPassInfo {
     // generated part
