+++ cocos/core/renderer/core/texture-buffer-pool.ts
@@ -3,8 +3,8 @@
  * @hidden
  */
 
-import { getTypedArrayConstructor, GFXBufferTextureCopy, GFXFormat, GFXFormatInfos, GFXTextureType, GFXTextureUsageBit, GFXTextureFlagBit } from '../../gfx/define';
-import { GFXTexture, GFXTextureInfo, GFXDevice } from '../../gfx';
+import { getTypedArrayConstructor, GFXFormat, GFXFormatInfos, GFXTextureType, GFXTextureUsageBit } from '../../gfx/define';
+import { GFXTexture, GFXTextureInfo, GFXDevice, GFXBufferTextureCopy } from '../../gfx';
 
 export function nearestPOT (num: number): number {
     --num;
