+++ cocos/core/utils/profiler/profiler.ts
@@ -28,8 +28,8 @@ import { MeshRenderer } from '../../3d/framework/mesh-renderer';
 import { Camera } from '../../3d/framework/camera-component';
 import { createMesh } from '../../3d/misc/utils';
 import { Material } from '../../assets/material';
-import { GFXBufferTextureCopy, GFXClearFlag, GFXFormat, GFXTextureType, GFXTextureUsageBit } from '../../gfx/define';
-import { GFXTexture, GFXTextureInfo, GFXDevice } from '../../gfx';
+import { GFXClearFlag, GFXFormat, GFXTextureType, GFXTextureUsageBit } from '../../gfx/define';
+import { GFXTexture, GFXTextureInfo, GFXDevice, GFXBufferTextureCopy } from '../../gfx';
 import { Vec4 } from '../../math';
 import { Layers } from '../../scene-graph';
 import { Node } from '../../scene-graph/node';
